py_db/requirements.txt
kafka-python
mysql-connector-python
